USER MANAGEMENT/Logout Session/requirements.txt
Flask
Werkzeug
flask-swagger-ui
flask-cors
python-dotenv